XW-4523: XmlHelper class
