+++ includes/wikia/helpers/tests/HtmlHelperTest.php
@@ -120,31 +120,4 @@ class HtmlHelperTest extends TestCase {
 			]
 		];
 	}
-
-	/**
-	 * @dataProvider renameNodeProvider
-	 */
-	public function testRenameNode( $html, $id, $expected, $message ) {
-		$document = HtmlHelper::createDOMDocumentFromText( $html );
-		HtmlHelper::renameNode( $document->getElementById( $id ), 'span' );
-
-		$this->assertEquals( $expected, HtmlHelper::getBodyHtml( $document ), $message );
-	}
-
-	public function renameNodeProvider() {
-		return [
-			[
-				'html' => '<p id="rename"><a href="http://example.com"></a></p>',
-				'id' => 'rename',
-				'expected' => '<span id="rename"><a href="http://example.com"></a></span>',
-				'message' => 'top level is not renamed'
-			],
-			[
-				'html' => '<div><h2 id="rename"><a href="http://example.com"></a></h2></p>',
-				'id' => 'rename',
-				'expected' => '<div><span id="rename"><a href="http://example.com"></a></span></div>',
-				'message' => 'nested node is not renamed'
-			]
-		];
-	}
 }
